# backend/src/vouchers/dto/update-voucher.dto.ts
import { PartialType, OmitType } from '@nestjs/swagger';
import { CreateVoucherDto } from './create-voucher.dto';

// Cannot update voucher type after creation
export class UpdateVoucherDto extends PartialType(
  OmitType(CreateVoucherDto, ['voucherType'] as const),
) {}