# old-version/modernization-design/api_spec.yaml
openapi: 3.0.3
info:
  title: Advance ERP API
  description: |
    Modern REST API for Advance ERP System
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Pagination, filtering, and sorting
    - Comprehensive error handling
    - Audit trails for all operations
    
    ## Authentication
    All endpoints (except `/auth/login`) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```
    
  version: 1.0.0
  contact:
    name: ERP Support
    email: support@advanceerp.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://staging-api.advanceerp.com/api
    description: Staging server
  - url: https://api.advanceerp.com/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 3600
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  
  # ============================================================================
  # ACCOUNTS (CHART OF ACCOUNTS)
  # ============================================================================
  /accounts:
    get:
      tags: [Accounts]
      summary: List all accounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category
          in: query
          schema:
            type: string
            enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        - name: accountType
          in: query
          schema:
            type: string
            enum: [CONTROL, SUB_CONTROL, DETAIL]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
          description: Search by code or name
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Accounts]
      summary: Create new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountDto'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    patch:
      tags: [Accounts]
      summary: Update account
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    delete:
      tags: [Accounts]
      summary: Delete account
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Account deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /vouchers:
    get:
      tags: [Vouchers]
      summary: List vouchers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: voucherType
          in: query
          schema:
            type: string
            enum: [JOURNAL, PAYMENT, RECEIPT]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: isPosted
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of vouchers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Voucher'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Vouchers]
      summary: Create voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoucherDto'
      responses:
        '201':
          description: Voucher created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /vouchers/{id}:
    get:
      tags: [Vouchers]
      summary: Get voucher by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Voucher details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
    
    patch:
      tags: [Vouchers]
      summary: Update voucher (only if not posted)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoucherDto'
      responses:
        '200':
          description: Voucher updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
    
    delete:
      tags: [Vouchers]
      summary: Delete voucher (only if not posted)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Voucher deleted
  /vouchers/{id}/post:
    post:
      tags: [Vouchers]
      summary: Post voucher
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Voucher posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
  /vouchers/{id}/unpost:
    post:
      tags: [Vouchers]
      summary: Unpost voucher
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Voucher unposted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
  
  # ============================================================================
  # GRN (GOODS RECEIPT NOTES)
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    
    # ========================================================================
    # ACCOUNTS
    # ========================================================================
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        parentAccountId:
          type: string
          format: uuid
          nullable: true
        accountType:
          type: string
          enum: [CONTROL, SUB_CONTROL, DETAIL]
        nature:
          type: string
          enum: [DEBIT, CREDIT]
        category:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        isActive:
          type: boolean
        openingBalance:
          type: number
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateAccountDto:
      type: object
      required: [name, nature, category, accountType]
      properties:
        name:
          type: string
          maxLength: 200
        parentAccountId:
          type: string
          format: uuid
        accountType:
          type: string
          enum: [CONTROL, SUB_CONTROL, DETAIL]
        nature:
          type: string
          enum: [DEBIT, CREDIT]
        category:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        openingBalance:
          type: number
          default: 0
        metadata:
          type: object
    UpdateAccountDto:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        metadata:
          type: object
    
    # ========================================================================
    # VOUCHERS
    # ========================================================================
    Voucher:
      type: object
      properties:
        id:
          type: string
          format: uuid
        voucherNumber:
          type: string
        voucherType:
          type: string
          enum: [JOURNAL, PAYMENT, RECEIPT]
        voucherDate:
          type: string
          format: date
        description:
          type: string
        totalAmount:
          type: number
        isPosted:
          type: boolean
        details:
          type: array
          items:
            $ref: '#/components/schemas/VoucherDetail'
        createdAt:
          type: string
          format: date-time
    VoucherDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountCode:
          type: string
        accountName:
          type: string
        description:
          type: string
        debitAmount:
          type: number
        creditAmount:
          type: number
        lineNumber:
          type: integer
    CreateVoucherDto:
      type: object
      required: [voucherType, voucherDate, details]
      properties:
        voucherType:
          type: string
          enum: [JOURNAL, PAYMENT, RECEIPT]
        voucherDate:
          type: string
          format: date
        description:
          type: string
        details:
          type: array
          minItems: 2
          items:
            type: object
            required: [accountCode, debitAmount, creditAmount]
            properties:
              accountCode:
                type: string
              description:
                type: string
              debitAmount:
                type: number
                minimum: 0
              creditAmount:
                type: number
                minimum: 0
    UpdateVoucherDto:
      type: object
      properties:
        description:
          type: string
        details:
          type: array
          items:
            type: object
    
    # ========================================================================
    # GRN
    # ========================================================================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 400
            message: Validation failed
            error: Bad Request
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 401
            message: Invalid credentials
            error: Unauthorized
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 404
            message: Resource not found
            error: Not Found